Group `PasswordGenereted` module into its own modules
