+++ KrakenVault/Password Generated/Domain/PasswordComponents.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright © 2022 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+struct PasswordComponents {
+    static let alphabet = ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z"]
+    static let specialCharactersArray = ["(", ")", "{", "}", "[", "]", "/", "+", "*", "$", ">", ".", "|", "^", "?", "&"]
+    static let numbersArray = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
+}
