+++ KrakenVault/Password Generated/PasswordGeneratedAction.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © 2022 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+enum PasswordGeneratedAction {
+    case generate
+}
